+++ .gitignore
@@ -15,3 +15,13 @@ rusty-tags.*
 #Eclipse project files
 .cproject
 .project
+
+# Compiled python modules.
+*.pyc
+
+# Setuptools distribution folder.
+/scripts/dist/
+
+# Python egg metadata, regenerated from source files by setuptools.
+/scripts/*.egg-info
+/scripts/*.egg
